+++ advinst-agent/src/main/java/jetbrains/buildServer/advinst/agent/AdvinstAgentListener.java
@@ -11,31 +11,26 @@ import jetbrains.buildServer.util.positioning.PositionConstraint;
 import org.jetbrains.annotations.NotNull;
 
 
-public class AdvinstAgentListener extends AgentLifeCycleAdapter implements PositionAware
-{
+public class AdvinstAgentListener extends AgentLifeCycleAdapter implements PositionAware {
 
   BuildAgentConfiguration mAgentConfiguration;
 
-  public AdvinstAgentListener(@NotNull final BuildAgentConfiguration agentConfiguration, @NotNull final EventDispatcher<AgentLifeCycleListener> events)
-  {
+  public AdvinstAgentListener(@NotNull final BuildAgentConfiguration agentConfiguration, @NotNull final EventDispatcher<AgentLifeCycleListener> events) {
     this.mAgentConfiguration = agentConfiguration;
     events.addListener(this);
   }
 
   @NotNull
-  public String getOrderId()
-  {
+  public String getOrderId() {
     return AdvinstConstants.RUNNER_TYPE;
   }
 
   @NotNull
-  public PositionConstraint getConstraint()
-  {
+  public PositionConstraint getConstraint() {
     return PositionConstraint.last();
   }
 
   @Override
-  public void agentInitialized(@NotNull final BuildAgent agent)
-  {
+  public void agentInitialized(@NotNull final BuildAgent agent) {
   }
 }