+++ advinst-common/src/main/java/jetbrains/buildServer/advinst/common/AdvinstAipReader.java
@@ -11,6 +11,7 @@ import javax.xml.xpath.XPath;
 import javax.xml.xpath.XPathConstants;
 import javax.xml.xpath.XPathExpressionException;
 import javax.xml.xpath.XPathFactory;
+
 import org.w3c.dom.Attr;
 import org.w3c.dom.Document;
 import org.w3c.dom.Node;
@@ -22,8 +23,7 @@ import org.xml.sax.SAXException;
  *
  * @author Ciprian Burca
  */
-public class AdvinstAipReader
-{
+public class AdvinstAipReader {
 
   private final String mAipFile;
   Document mXmlDocument = null;
@@ -33,8 +33,7 @@ public class AdvinstAipReader
    *
    * @param aAipFile Path to Advanced Installer project file (.AIP)
    */
-  public AdvinstAipReader(String aAipFile)
-  {
+  public AdvinstAipReader(String aAipFile) {
     this.mAipFile = aAipFile;
   }
 
@@ -44,8 +43,7 @@ public class AdvinstAipReader
    * @return strings list containing the build names
    * @throws AdvinstException
    */
-  public List<String> getBuilds() throws AdvinstException
-  {
+  public List<String> getBuilds() throws AdvinstException {
     List<String> aipBuilds = new ArrayList<String>();
 
     loadXmlFile();
@@ -53,20 +51,15 @@ public class AdvinstAipReader
     final String buildsXPAth = "/DOCUMENT/COMPONENT[@cid='caphyon.advinst.msicomp.BuildComponent']/ROW";
     XPath xPath = XPathFactory.newInstance().newXPath();
     NodeList buildRows;
-    try
-    {
+    try {
       buildRows = (NodeList) xPath.evaluate(buildsXPAth, mXmlDocument, XPathConstants.NODESET);
-    }
-    catch (XPathExpressionException ex)
-    {
+    } catch (XPathExpressionException ex) {
       throw new AdvinstException("Failed read AIP builds. Exception: " + ex.getMessage(), ex);
     }
-    for (int i = 0; i < buildRows.getLength(); i++)
-    {
+    for (int i = 0; i < buildRows.getLength(); i++) {
       Node buildRow = buildRows.item(i);
       Attr nameAttr = (Attr) buildRow.getAttributes().getNamedItem("BuildName");
-      if (null != nameAttr)
-      {
+      if (null != nameAttr) {
         aipBuilds.add(nameAttr.getValue());
       }
     }
@@ -74,52 +67,39 @@ public class AdvinstAipReader
     return aipBuilds;
   }
 
-  public boolean isValidAip() throws AdvinstException
-  {
+  public boolean isValidAip() throws AdvinstException {
     loadXmlFile();
     NodeList children = mXmlDocument.getChildNodes();
-    if (children.getLength() != 1)
-    {
+    if (children.getLength() != 1) {
       return false;
     }
 
-    if (!"DOCUMENT".equals(children.item(1).getNodeName()))
-    {
+    if (!"DOCUMENT".equals(children.item(1).getNodeName())) {
       return false;
     }
 
     Attr nameAttr = (Attr) children.item(1).getAttributes().getNamedItem("Type");
-    if (null == nameAttr)
-    {
+    if (null == nameAttr) {
       return false;
     }
 
     return true;
   }
 
-  private void loadXmlFile() throws AdvinstException
-  {
-    try
-    {
-      if (null != mXmlDocument)
-      {
+  private void loadXmlFile() throws AdvinstException {
+    try {
+      if (null != mXmlDocument) {
         return;
       }
 
       DocumentBuilderFactory documentBuilderFactory = DocumentBuilderFactory.newInstance();
       DocumentBuilder documentBuilder = documentBuilderFactory.newDocumentBuilder();
       mXmlDocument = documentBuilder.parse(new File(mAipFile));
-    }
-    catch (SAXException ex)
-    {
+    } catch (SAXException ex) {
       throw new AdvinstException("Failed to load AIP file. Exception: " + ex.getMessage(), ex);
-    }
-    catch (ParserConfigurationException ex)
-    {
+    } catch (ParserConfigurationException ex) {
       throw new AdvinstException("Failed to load AIP file. Exception: " + ex.getMessage(), ex);
-    }
-    catch (IOException ex)
-    {
+    } catch (IOException ex) {
       throw new AdvinstException("Failed to load AIP file. Exception: " + ex.getMessage(), ex);
     }
   }