+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Changelog of Advanced Installer plugin for TeamCity.
 
+### Advanced Installer TeamCity Plugin  1.2
+
+__*The minimum required TeamCity version is now 2017.1.*__
+
+Features:
+* Automated deploy of Advanced Installer tool on agents.
+
 ### Advanced Installer TeamCity Plugin  1.1.2
 
 Bugs: